refactor  : refactoring the dynamic blog flow and breaking code into components.
