+++ src/types/blog.ts
@@ -1,0 +1,5 @@
+export interface BlogParams {
+  title: string;
+  slug: string;
+  imageUrl: string;
+}
